+++ org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/lookup/TypeConstants.java
@@ -138,11 +138,12 @@ public interface TypeConstants {
 	char[] NON_SEALED = "non-sealed".toCharArray(); //$NON-NLS-1$
 	String KEYWORD_EXTENDS = "extends"; //$NON-NLS-1$
 	String IMPLEMENTS = "implements"; //$NON-NLS-1$
+	// JEP 401 Value Classes and Objects
+	char[] VALUE_RESTRICTED_KEYWORD = "value".toCharArray(); //$NON-NLS-1$
 
     // jsr308
 	char[] TYPE_USE_TARGET  = "TYPE_USE".toCharArray(); //$NON-NLS-1$
 	char[] TYPE_PARAMETER_TARGET = "TYPE_PARAMETER".toCharArray(); //$NON-NLS-1$
-	char[] VALUE_RESTRICTED_KEYWORD = "value".toCharArray(); //$NON-NLS-1$
 
     // common 3rd party package components:
     char[] ORG = "org".toCharArray(); //$NON-NLS-1$
