+++ org.eclipse.jdt.core.tests.compiler/src/org/eclipse/jdt/core/tests/compiler/regression/ValueClassesAndObjectsTest.java
@@ -15,22 +15,17 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 package org.eclipse.jdt.core.tests.compiler.regression;
-import java.util.Map;
 import junit.framework.Test;
 import org.eclipse.jdt.core.tests.util.PreviewTest;
-import org.eclipse.jdt.internal.compiler.batch.FileSystem;
-import org.eclipse.jdt.internal.compiler.env.INameEnvironment;
-import org.eclipse.jdt.internal.compiler.impl.CompilerOptions;
+
 @PreviewTest
-public class ValueClassesAndObjectsTest extends AbstractRegressionTest9 {
+public class ValueClassesAndObjectsTest extends AbstractRegressionTestCommon {
 	private static final JavacTestOptions JAVAC_OPTIONS = new JavacTestOptions("--enable-preview -source 26");
-	private static final String[] VMARGS = new String[] {"--enable-preview"};
 	static {
 //		TESTS_NUMBERS = new int [] { 1 };
 //		TESTS_RANGE = new int[] { 1, -1 };
 //		TESTS_NAMES = new String[] { "testIssue3536" };
 	}
-	private String extraLibPath;
 	public static Class<?> testClass() {
 		return ValueClassesAndObjectsTest.class;
 	}
@@ -51,101 +46,12 @@ public class ValueClassesAndObjectsTest extends AbstractRegressionTest9 {
 		super.tearDown();
 		this.runJavacOptIn = false; // do it last, so super can still clean up
 	}
+
+	@Override
+	protected JavacTestOptions getJavacTestOptions() {
+		return JAVAC_OPTIONS;
+	}
 	// =================================================
-	// Enables the tests to run individually
-	protected Map<String, String> getCompilerOptions(boolean preview) {
-		Map<String, String> defaultOptions = super.getCompilerOptions();
-		defaultOptions.put(CompilerOptions.OPTION_Compliance, CompilerOptions.VERSION_26);
-		defaultOptions.put(CompilerOptions.OPTION_Source, CompilerOptions.VERSION_26);
-		defaultOptions.put(CompilerOptions.OPTION_TargetPlatform, CompilerOptions.VERSION_26);
-		defaultOptions.put(CompilerOptions.OPTION_EnablePreviews, preview ? CompilerOptions.ENABLED : CompilerOptions.DISABLED);
-		defaultOptions.put(CompilerOptions.OPTION_ReportPreviewFeatures, CompilerOptions.WARNING);
-		return defaultOptions;
-	}
-	protected Map<String, String> getCompilerOptions() {
-		return getCompilerOptions(false);
-	}
-	protected String[] getDefaultClassPaths() {
-		String[] libs = DefaultJavaRuntimeEnvironment.getDefaultClassPaths();
-		if (this.extraLibPath != null) {
-			String[] l = new String[libs.length + 1];
-			System.arraycopy(libs, 0, l, 0, libs.length);
-			l[libs.length] = this.extraLibPath;
-			return l;
-		}
-		return libs;
-	}
-	@Override
-	protected INameEnvironment getNameEnvironment(final String[] testFiles, String[] classPaths, Map<String, String> options) {
-		this.classpaths = classPaths == null ? getDefaultClassPaths() : classPaths;
-		INameEnvironment[] classLibs = getClassLibs(false, options);
-		for (INameEnvironment nameEnvironment : classLibs) {
-			((FileSystem) nameEnvironment).scanForModules(createParser());
-		}
-		return new InMemoryNameEnvironment9(testFiles, this.moduleMap, classLibs);
-	}
-	@Override
-	protected void runConformTest(String[] testFiles, String expectedOutput) {
-		runConformTest(testFiles, expectedOutput, getCompilerOptions(true), VMARGS, JAVAC_OPTIONS);
-	}
-	@Override
-	protected void runConformTest(String[] testFiles, String expectedOutput, Map<String, String> customOptions) {
-		if(!isJRE26Plus)
-			return;
-		runConformTest(testFiles, expectedOutput, customOptions, VMARGS, JAVAC_OPTIONS);
-	}
-	protected void runConformTest(
-			String[] testFiles,
-			String expectedOutputString,
-			String[] classLibraries,
-			boolean shouldFlushOutputDirectory,
-			String[] vmArguments) {
-			runTest(
-		 		// test directory preparation
-				shouldFlushOutputDirectory /* should flush output directory */,
-				testFiles /* test files */,
-				// compiler options
-				classLibraries /* class libraries */,
-				null /* no custom options */,
-				false /* do not perform statements recovery */,
-				null /* no custom requestor */,
-				// compiler results
-				false /* expecting no compiler errors */,
-				null /* do not check compiler log */,
-				// runtime options
-				false /* do not force execution */,
-				vmArguments /* vm arguments */,
-				// runtime results
-				expectedOutputString /* expected output string */,
-				null /* do not check error string */,
-				// javac options
-				JavacTestOptions.DEFAULT /* default javac test options */);
-		}
-	protected void runNegativeTest(String[] testFiles, String expectedCompilerLog) {
-		Map<String, String> customOptions = getCompilerOptions(true);
-		Runner runner = new Runner();
-		runner.testFiles = testFiles;
-		runner.expectedCompilerLog = expectedCompilerLog;
-		runner.javacTestOptions = JAVAC_OPTIONS;
-		runner.customOptions = customOptions;
-		runner.expectedJavacOutputString = null;
-		runner.runNegativeTest();
-	}
-	protected void runNegativeTest(
-			String[] testFiles,
-			String expectedCompilerLog,
-			String javacLog,
-			String[] classLibraries,
-			boolean shouldFlushOutputDirectory,
-			Map<String, String> customOptions) {
-		Runner runner = new Runner();
-		runner.testFiles = testFiles;
-		runner.expectedCompilerLog = expectedCompilerLog;
-		runner.javacTestOptions = JAVAC_OPTIONS;
-		runner.customOptions = customOptions;
-		runner.expectedJavacOutputString = javacLog;
-		runner.runNegativeTest();
-	}
 	// https://cr.openjdk.org/~dlsmith/jep401/latest/
   public void testValueTypes_001() {
 		runNegativeTest(new String[] {
